Implement all for all datasets

+++ nets/load_net.py
@@ -12,4 +12,4 @@ def GraphTransformer(net_params):
 def gnn_model(MODEL_NAME, net_params):
     models = {"GraphTransformer": GraphTransformer}
 
-    return models[MODEL_NAME](net_params)
+    return models[MODEL_NAME](net_params).float()
